geo functions removed from map
