+++ geo/map_processing/__init__.py
@@ -1,1 +1,2 @@
-from .mm import Map
+from .map_structure import Map
+import .geo_utils
